bios: ebda_region: avoid warnings on 64 bit build.

+++ src/bios/ebda_region/ebda_region.c
@@ -118,8 +118,8 @@ static int ebda_test1(fwts_framework *fw)
 				if (start_addr <= ebda_addr && end_addr > ebda_addr) {
 					fwts_passed(fw, "EBDA region mapped at 0x%lx and reserved as a %lldK region in E820 table at 0x%llx..0x%llx.",
 						ebda_addr, 
-						(end_addr - start_addr) / 1024,
-						start_addr, end_addr);
+						(unsigned long long int)(end_addr - start_addr) / 1024,
+						(unsigned long long int)start_addr, (unsigned long long int)end_addr);
 					passed = 1;
 					break;
 				}
